\name{AE-package}
\alias{AE-package}
\alias{AE}
\docType{package}
\title{
\packageTitle{AE}
}
\description{
\packageDescription{AE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AE}
\packageIndices{AE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{AE}

Maintainer: \packageMaintainer{AE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
