% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae.R
\name{ae}
\alias{ae}
\title{Attituduinal Equity}
\usage{
ae(data, id, relevance, closeness)
}
\arguments{
\item{data}{A data frame.}

\item{id}{name of id variable, within "".}

\item{relevance}{transposed relevance variable, within "".}

\item{closeness}{transposed closeness variable, within "".}
}
\value{
Returns data frame added with avg brand score and Attitudinal Equity.
}
\description{
Function computes the avg brand score and Attitudinal Equity. Data should be structured in long format.
}
\references{
\url{https://www.researchgate.net/publication/252568552_A_new_measure_of_brand_attitudinal_equity_based_on_the_Zipf_distribution}
}
