\name{ae-package}
\alias{ae-package}
\alias{ae}
\docType{package}
\title{
\packageTitle{ae}
}
\description{
\packageDescription{ae}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ae}
\packageIndices{ae}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ae}

Maintainer: \packageMaintainer{ae}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
